third_party/blink/web_tests/platform/generic/external/wpt/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/legend-grid-flex-multicol-expected.txt

This is a testharness.js-based test.
PASS flex
FAIL inline-flex assert_equals: display expected "inline-flex" but got "flex"
PASS grid
FAIL inline-grid assert_equals: display expected "inline-grid" but got "grid"
PASS multicol
Harness: the test ran to completion.
